exercise list changes
