+++ src/components/Exercises.tsx
@@ -4,6 +4,7 @@ import { Exercise } from "../models/exercise"
 import { getExercisesByUserId } from "../services/apiService";
 import { decodeToken } from "../services/tokenService";
 import { ExerciseById } from "./Exercise";
+import { AddExercise } from "./AddExercise";
 
 interface Props {
 }
@@ -11,11 +12,18 @@ interface Props {
 export const Exercises = (props: Props) => {
     const { auth } = useContext(AuthContext);
     const [exercises, setExercises] = useState<Exercise[]>([])
+    const [addExercise, setAddExercise] = useState(false)
+    const buttonText = addExercise ? "Cancel" : "Add Exercise"
 
-    const handleChange = async () => {
+    const closeAddExerciseModal = (): void => {
+        setAddExercise(false);
+    }
+
+    const triggerExercisesListChange = async () => {
         const { id } = decodeToken(auth.token)
         const exercises = await getExercisesByUserId(id)
         setExercises(exercises)
+        console.log('triggerExerciseListChange()')
     }
 
     useEffect(() => {
@@ -27,13 +35,28 @@ export const Exercises = (props: Props) => {
         fetchAndSetExercises()
     }, [auth.token])
 
+    if (addExercise === true) {
+        return (
+            <div className="exercise-list">
+                <ul>
+                    {exercises.map(exercise => (
+                        <ExerciseById key={exercise.id} id={exercise.id} name={exercise.name} max={exercise.max} units={exercise.units} triggerExercisesChange={triggerExercisesListChange} />
+                    ))}
+                </ul>
+                <button onClick={() => setAddExercise(!addExercise)}>{buttonText}</button>
+                <AddExercise closeAddExerciseModal={closeAddExerciseModal} triggerExercisesChange={triggerExercisesListChange} />
+            </div>
+        )
+    }
+
     return (
         <div className="exercise-list">
             <ul>
                 {exercises.map(exercise => (
-                    <ExerciseById key={exercise.id} id={exercise.id} name={exercise.name} max={exercise.max} units={exercise.units} handler={handleChange} />
+                    <ExerciseById key={exercise.id} id={exercise.id} name={exercise.name} max={exercise.max} units={exercise.units} triggerExercisesChange={triggerExercisesListChange} />
                 ))}
             </ul>
+            <button onClick={() => setAddExercise(!addExercise)}>{buttonText}</button>
         </div>
     )
 }
